+++ cypress/integration/native-uploady/NativeUploady-simple-spec.js
@@ -1,16 +1,19 @@
 import uploadFile from "../uploadFile";
-import { WAIT_SHORT } from "../../constants";
 
 describe("NativeUploady - Simple", () => {
     const fileName = "flower.jpg";
 
     before(() => {
-        cy.visitStory("nativeUploady", "simple&knob-mock send delay_Upload Destination=100");
+        cy.visitStory(
+            "nativeUploady",
+            "simple",
+            { mockDelay: 100 }
+        );
     });
 
     it("should use native uploady", () => {
         uploadFile(fileName, () => {
-            cy.wait(WAIT_SHORT);
+            cy.waitExtraShort();
             cy.storyLog().assertFileItemStartFinish(fileName, 1);
         });
     });