+++ cypress/integration/upload-button/UploadButton-differentConfig-spec.js
@@ -1,5 +1,4 @@
 import uploadFile from "../uploadFile";
-import { WAIT_X_SHORT } from "../../constants";
 
 describe("UploadButton - Different Configuration", () => {
     const fileName = "flower.jpg";
@@ -11,13 +10,13 @@ describe("UploadButton - Different Configuration", () => {
     it("should allow overriding upload options from button", () => {
         //test button with autoUpload = false
         uploadFile(fileName, () => {
-            cy.wait(WAIT_X_SHORT);
+            cy.waitExtraShort();
             cy.storyLog().assertLogEntryCount(1);
         }, "#upload-a");
 
         //test other button with custom destination header
         uploadFile(fileName, () => {
-            cy.wait(WAIT_X_SHORT);
+            cy.waitExtraShort();
 
             cy.storyLog().assertLogEntryContains(1, {
                 destination: {