+++ cypress/integration/upload-button/UploadButton-group-spec.js
@@ -1,17 +1,22 @@
 import intercept from "../intercept";
 import { uploadFileTimes } from "../uploadFile";
-import { ITEM_START, ITEM_FINISH, UPLOAD_URL } from "../../constants";
-import { WAIT_MEDIUM } from "../../constants";
+import { ITEM_START, ITEM_FINISH } from "../../constants";
 
 describe("UploadButton - Simple - Multiple files", () => {
     const fileName = "flower.jpg";
 
     beforeEach(() => {
-        cy.visitStory("uploadButton", "simple&knob-destination_Upload Destination=local&knob-group files in single request_Upload Settings=true&knob-max in group_Upload Settings=2");
+        cy.visitStory(
+            "uploadButton",
+            "simple",
+            {
+                useMock: false,
+                grouped: true,
+            });
     });
 
     it("should show upload preview for multiple files", () => {
-        intercept(UPLOAD_URL);
+        intercept();
 
         uploadFileTimes(fileName, () => {
             cy.wait("@uploadReq")
@@ -25,7 +30,7 @@ describe("UploadButton - Simple - Multiple files", () => {
                     expect(formData["file"]).to.eq("flower3.jpg");
                 });
 
-            cy.wait(WAIT_MEDIUM);
+            cy.waitMedium();
 
             cy.storyLog().assertLogPattern(ITEM_START, { index: 1 });
             cy.storyLog().assertLogPattern(ITEM_START, { index: 2 });