+++ cypress/integration/upload-drop-zone/UploadDropZone-handle-only-files-spec.js
@@ -1,11 +1,15 @@
-import { ITEM_START, WAIT_X_SHORT } from "../../constants";
+import { ITEM_START } from "../../constants";
 import dropFile from "../dropFile";
 
 describe("UploadDropZone - shouldHandleDrag for files only", () => {
     const fileName = "flower.jpg";
 
     before(() => {
-        cy.visitStory("uploadDropZone", "with-full-screen&knob-mock send delay_Upload Destination=100");
+        cy.visitStory(
+            "uploadDropZone",
+            "with-full-screen",
+            { mockDelay: 100 }
+        );
     });
 
     it("should not show drag overlay if not all files", () => {
@@ -26,7 +30,7 @@ describe("UploadDropZone - shouldHandleDrag for files only", () => {
             .trigger("dragleave", {});
 
         dropFile(fileName, () => {
-            cy.wait(WAIT_X_SHORT);
+            cy.waitExtraShort();
             cy.storyLog().assertLogPattern(ITEM_START);
         });
     });