+++ cypress/integration/upload-drop-zone/UploadDropZone-simple-spec.js
@@ -1,16 +1,15 @@
 import dropFile from "../dropFile";
-import { WAIT_X_SHORT } from "../../constants";
 
 describe("UploadDropZone - Simple", () => {
     const fileName = "flower.jpg";
 
     before(() => {
-        cy.visitStory("uploadDropZone", "simple&knob-mock send delay_Upload Destination=100");
+        cy.visitStory("uploadDropZone", "simple", { mockDelay: 100 });
     });
 
     it("should upload dropped file", () => {
         dropFile(fileName, () => {
-            cy.wait(WAIT_X_SHORT);
+            cy.waitExtraShort();
             cy.storyLog().assertFileItemStartFinish(fileName, 1);
         });
     });