+++ cypress/integration/uploady/Uploady-invalid-request-pre-send-spec.js
@@ -1,4 +1,4 @@
-import { ITEM_ERROR, ITEM_FINISH, ITEM_START, UPLOAD_URL } from "../../constants";
+import { ITEM_ERROR, ITEM_FINISH, ITEM_START } from "../../constants";
 import intercept from "../intercept";
 import { uploadFileTimes } from "../uploadFile";
 
@@ -6,11 +6,15 @@ describe("Uploady - invalid requestPreSend", () => {
     const fileName = "flower.jpg";
 
     beforeEach(() => {
-        cy.visitStory("uploady", "test-invalid-pre-send&knob-destination_Upload Destination=local&knob-mock send delay_Upload Destination=1000&knob-multiple files_Upload Settings=true&knob-group files in single request_Upload Settings=&knob-max in group_Upload Settings=2&knob-auto upload on add_Upload Settings=true");
+        cy.visitStory(
+            "uploady",
+            "test-invalid-pre-send",
+            { useMock: false }
+        );
     });
 
     it("should fail invalid updated data from pre send - forbidden item props", () => {
-        intercept(UPLOAD_URL);
+        intercept();
 
         uploadFileTimes(fileName, () => {
             cy.waitShort();