+++ packages/ui/tus-uploady/TusUploady.stories.js
@@ -1,17 +1,16 @@
 // @flow
 import React, { useState } from "react";
 import UploadButton from "@rpldy/upload-button";
-import { number, boolean } from "@storybook/addon-knobs";
 import {
-    DESTINATION_TYPES,
+    DEFAULT_CHUNK_SIZE,
     getCsfExport,
-    KNOB_GROUPS,
-    useStoryUploadySetup,
+    createUploadyStory,
+    getTusDestinationOptions,
     type CsfExport,
+    type UploadyStory,
 } from "../../../story-helpers";
 import retryEnhancer, { useRetry, useRetryListener } from "@rpldy/retry-hooks";
-import TusUploady,
-{
+import TusUploady, {
     useAbortAll,
     useItemProgressListener,
     useItemStartListener,
@@ -23,85 +22,86 @@ import TusUploady,
     useTusResumeStartListener,
 } from "./src";
 
-import readme from "./README.md";
-
+import Readme from "./TusUploady.storydoc.mdx";
 import type { Node } from "react";
-
-const useTusStoryHelper = () => {
-	const setup = useStoryUploadySetup({
-		noGroup: true,
-		destinations: [DESTINATION_TYPES.url, DESTINATION_TYPES.local],
-		noLong: true,
-	});
-
-	const chunkSize = number("chunk size (bytes)", 524288, {}, KNOB_GROUPS.SETTINGS);
-	const forgetOnSuccess = boolean("forget on success", false, KNOB_GROUPS.SETTINGS);
-	const resume = boolean("enable resume (storage)", true, KNOB_GROUPS.SETTINGS);
-	const ignoreModifiedDateInStorage = boolean("ignore modifiedDate in resume storage", false, KNOB_GROUPS.SETTINGS);
-	const sendDataOnCreate = boolean("send data on create", false, KNOB_GROUPS.SETTINGS);
-	const sendWithCustomHeader = boolean("send custom header", false, KNOB_GROUPS.SETTINGS);
-
-	return {
-		...setup,
-		chunkSize,
-		forgetOnSuccess,
-		resume,
-		ignoreModifiedDateInStorage,
-		sendDataOnCreate,
-		sendWithCustomHeader,
-	};
-};
 
 const AbortButton = () => {
     const abortAll = useAbortAll();
 
-    return <button id="abort-btn" onClick={abortAll}>Abort</button>;
+    return (
+        <button id="abort-btn" onClick={abortAll}>
+            Abort
+        </button>
+    );
 };
 
 const ItemProgress = () => {
     const [progress, setProgress] = useState<{ id: string, text: string }[]>([]);
 
     useItemProgressListener(({ id, batchId, loaded, completed }) => {
-        setProgress((latest) => latest.concat({ id: batchId + id + loaded, text: `${id}: LOADED - ${loaded} - COMPLETED - ${completed}` }));
+        setProgress((latest) =>
+            latest.concat({
+                id: batchId + id + loaded,
+                text: `${id}: LOADED - ${loaded} - COMPLETED - ${completed}`,
+            }),
+        );
     });
 
     useItemAbortListener(({ id, batchId }) => {
-        setProgress((latest) => latest.concat({ id: batchId + id + "_abort", text: `${id}: ABORT !!!!!! ----------` }));
+        setProgress((latest) =>
+            latest.concat({
+                id: batchId + id + "_abort",
+                text: `${id}: ABORT !!!!!! ----------`,
+            }),
+        );
     });
 
     return (
         <>
-            {progress.map((p) => <p key={p.id} data-id={p.id}>{p.text}</p>)}
+            {progress.map((p) => (
+                <p key={p.id} data-id={p.id}>
+                    {p.text}
+                </p>
+            ))}
             <br/>
             <button onClick={() => setProgress([])}>clear progress</button>
         </>
     );
 };
 
-export const Simple = (): Node => {
-	const storySetup = useTusStoryHelper();
-	let { destination } = storySetup;
-	const { enhancer, chunkSize, forgetOnSuccess, resume, ignoreModifiedDateInStorage, sendDataOnCreate, sendWithCustomHeader } = storySetup;
+export const Simple: UploadyStory = createUploadyStory(
+    ({
+         destination,
+         enhancer,
+         chunkSize,
+         forgetOnSuccess,
+         resumeStorage,
+         ignoreModifiedDateInStorage,
+         sendDataOnCreate,
+         sendWithCustomHeader,
+     }): Node => {
+        const activeDestination = sendWithCustomHeader ?
+            { ...destination, headers: { "x-test": "abcd" } } :
+            destination
 
-	if (sendWithCustomHeader) {
-		destination = { ...destination, headers: { "x-test": "abcd" } };
-	}
-
-	return <TusUploady
-		debug
-		destination={destination}
-		enhancer={enhancer}
-		chunkSize={chunkSize}
-		forgetOnSuccess={forgetOnSuccess}
-		resume={resume}
-		ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
-		sendDataOnCreate={sendDataOnCreate}>
-		<UploadButton id="upload-button">Upload with TUS</UploadButton>
-        <br/>
-        <AbortButton/>
-        <ItemProgress/>
-	</TusUploady>;
-};
+        return (
+            <TusUploady
+                debug
+                destination={activeDestination}
+                enhancer={enhancer}
+                chunkSize={chunkSize}
+                forgetOnSuccess={forgetOnSuccess}
+                resume={resumeStorage}
+                ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
+                sendDataOnCreate={sendDataOnCreate}
+            >
+                <UploadButton id="upload-button">Upload with TUS</UploadButton>
+                <br/>
+                <AbortButton/>
+                <ItemProgress/>
+            </TusUploady>
+        );
+    });
 
 const DynamicUploadMeta = () => {
     useRequestPreSend(({ items }) => {
@@ -111,34 +111,42 @@ const DynamicUploadMeta = () => {
             options: {
                 params: {
                     "md-fileName": name,
-                }
-            }
-        }
+                },
+            },
+        };
     });
 
     return null;
 };
 
-export const WithDynamicMetadata = (): Node => {
-    const storySetup = useTusStoryHelper();
-    let { destination } = storySetup;
-    const { enhancer, chunkSize, forgetOnSuccess, resume, ignoreModifiedDateInStorage, sendDataOnCreate, sendWithCustomHeader } = storySetup;
-
-    return <TusUploady
-        debug
-        destination={destination}
-        enhancer={enhancer}
-        chunkSize={chunkSize}
-        forgetOnSuccess={forgetOnSuccess}
-        resume={resume}
-        ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
-        sendDataOnCreate={sendDataOnCreate}>
-        <UploadButton id="upload-button">Upload with TUS</UploadButton>
-        <br/>
-        <ItemProgress/>
-        <DynamicUploadMeta/>
-    </TusUploady>;
-};
+export const WithDynamicMetadata: UploadyStory = createUploadyStory(
+    ({
+         destination,
+         enhancer,
+         chunkSize,
+         forgetOnSuccess,
+         resumeStorage,
+         ignoreModifiedDateInStorage,
+         sendDataOnCreate,
+     }): Node => {
+        return (
+            <TusUploady
+                debug
+                destination={destination}
+                enhancer={enhancer}
+                chunkSize={chunkSize}
+                forgetOnSuccess={forgetOnSuccess}
+                resume={resumeStorage}
+                ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
+                sendDataOnCreate={sendDataOnCreate}
+            >
+                <UploadButton id="upload-button">Upload with TUS</UploadButton>
+                <br/>
+                <ItemProgress/>
+                <DynamicUploadMeta/>
+            </TusUploady>
+        );
+    });
 
 const TusConcatUploadLog = () => {
     const [log, setLog] = useState<string[]>([]);
@@ -151,27 +159,36 @@ const TusConcatUploadLog = () => {
         setLog((log) => log.concat("ITEM FINISHED UPLOADING!"));
     });
 
-    return (log.map((line) =>
-        <p key={line}>{line}</p>));
+    return log.map((line) => <p key={line}>{line}</p>);
 };
 
-export const WithTusConcatenation = (): Node => {
-	const { enhancer, destination, chunkSize, forgetOnSuccess, resume, ignoreModifiedDateInStorage, sendDataOnCreate } = useTusStoryHelper();
-
-	return <TusUploady
-		debug
-		destination={destination}
-		enhancer={enhancer}
-		chunkSize={chunkSize}
-		forgetOnSuccess={forgetOnSuccess}
-		resume={resume}
-		parallel={2}
-		ignoreModifiedDateInStorage={!ignoreModifiedDateInStorage}
-		sendDataOnCreate={sendDataOnCreate}>
-		<UploadButton>Upload with TUS Concatenation</UploadButton>
-        <TusConcatUploadLog/>
-	</TusUploady>;
-};
+export const WithTusConcatenation: UploadyStory = createUploadyStory(
+    ({
+         destination,
+         enhancer,
+         chunkSize,
+         forgetOnSuccess,
+         resumeStorage,
+         ignoreModifiedDateInStorage,
+         sendDataOnCreate,
+     }): Node => {
+        return (
+            <TusUploady
+                debug
+                destination={destination}
+                enhancer={enhancer}
+                chunkSize={chunkSize}
+                forgetOnSuccess={forgetOnSuccess}
+                resume={resumeStorage}
+                parallel={2}
+                ignoreModifiedDateInStorage={!ignoreModifiedDateInStorage}
+                sendDataOnCreate={sendDataOnCreate}
+            >
+                <UploadButton>Upload with TUS Concatenation</UploadButton>
+                <TusConcatUploadLog/>
+            </TusUploady>
+        );
+    });
 
 const RetryTus = () => {
     const retry = useRetry();
@@ -182,7 +199,7 @@ const RetryTus = () => {
     });
 
     useItemErrorListener(() => {
-       setShowRetry(true);
+        setShowRetry(true);
     });
 
     useRetryListener((items) => {
@@ -190,87 +207,131 @@ const RetryTus = () => {
     });
 
     return (
-        showRetry &&
-        <button
-            id="retry-tus-btn"
-            onClick={() => retry()}
-        >
-            Retry TUS
-        </button>
+        showRetry && (
+            <button id="retry-tus-btn" onClick={() => retry()}>
+                Retry TUS
+            </button>
+        )
     );
 };
 
-export const WithRetry = (): Node => {
-    const storySetup = useTusStoryHelper();
-    let { destination } = storySetup;
-    const { enhancer, chunkSize, forgetOnSuccess, resume, ignoreModifiedDateInStorage, sendDataOnCreate, sendWithCustomHeader } = storySetup;
+export const WithRetry: UploadyStory = createUploadyStory(
+    ({
+         destination,
+         enhancer,
+         chunkSize,
+         forgetOnSuccess,
+         resumeStorage,
+         ignoreModifiedDateInStorage,
+         sendDataOnCreate,
+         sendWithCustomHeader,
+     }): Node => {
+        const activeDestination = sendWithCustomHeader ?
+            { ...destination, headers: { "x-test": "abcd" } } :
+            destination;
 
-    if (sendWithCustomHeader) {
-        destination = { ...destination, headers: { "x-test": "abcd" } };
-    }
-
-    return <TusUploady
-        debug
-        destination={destination}
-        enhancer={composeEnhancers(enhancer, retryEnhancer)}
-        chunkSize={chunkSize}
-        forgetOnSuccess={forgetOnSuccess}
-        resume={resume}
-        ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
-        sendDataOnCreate={sendDataOnCreate}>
-        <UploadButton id="upload-button">Upload with TUS</UploadButton>
-        <br/>
-        <AbortButton/>
-        <RetryTus/>
-        <ItemProgress/>
-    </TusUploady>;
-};
+        return (
+            <TusUploady
+                debug
+                destination={activeDestination}
+                enhancer={composeEnhancers(enhancer, retryEnhancer)}
+                chunkSize={chunkSize}
+                forgetOnSuccess={forgetOnSuccess}
+                resume={resumeStorage}
+                ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
+                sendDataOnCreate={sendDataOnCreate}
+            >
+                <UploadButton id="upload-button">Upload with TUS</UploadButton>
+                <br/>
+                <AbortButton/>
+                <RetryTus/>
+                <ItemProgress/>
+            </TusUploady>
+        );
+    });
 
 const ResumeHandler = ({ cancelResume = false }: { cancelResume: boolean }) => {
     useTusResumeStartListener(() => {
         return cancelResume ? false : {
-            resumeHeaders: {
-                "x-another-header": "foo",
-                "x-test-override": "def"
-            }
-        }
+                resumeHeaders: {
+                    "x-another-header": "foo",
+                    "x-test-override": "def",
+                },
+            };
     });
 
     return null;
 };
 
-export const WithResumeStartHandler = (): Node => {
-    const storySetup = useTusStoryHelper();
-    let { destination } = storySetup;
-    const { enhancer, chunkSize, forgetOnSuccess, resume, ignoreModifiedDateInStorage, sendDataOnCreate, sendWithCustomHeader, extOptions } = storySetup;
+export const WithResumeStartHandler: UploadyStory = createUploadyStory(
+    ({
+         destination,
+         enhancer,
+         chunkSize,
+         forgetOnSuccess,
+         resumeStorage,
+         ignoreModifiedDateInStorage,
+         sendDataOnCreate,
+         sendWithCustomHeader,
+         extOptions,
+     }): Node => {
+        const activeDestination = sendWithCustomHeader ?
+            { ...destination, headers: { "x-test": "abcd" } } :
+            destination;
+        return (
+            <TusUploady
+                debug
+                destination={activeDestination}
+                enhancer={composeEnhancers(enhancer, retryEnhancer)}
+                chunkSize={chunkSize}
+                forgetOnSuccess={forgetOnSuccess}
+                resume={resumeStorage}
+                ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
+                sendDataOnCreate={sendDataOnCreate}
+                resumeHeaders={{
+                    "x-test-resume": "123",
+                    "x-test-override": "abc",
+                }}
+            >
+                <UploadButton id="upload-button">Upload with TUS</UploadButton>
+                <br/>
+                <AbortButton/>
+                <RetryTus/>
+                <ItemProgress/>
+                <ResumeHandler cancelResume={extOptions?.tusCancelResume}/>
+            </TusUploady>
+        );
+    });
 
-    if (sendWithCustomHeader) {
-        destination = { ...destination, headers: { "x-test": "abcd" } };
+const tusUploadyStories: CsfExport = getCsfExport(TusUploady, "Tus Uploady", Readme, {
+    pkg: "tus-uploady",
+    section: "UI",
+    parameters: {
+        controls: {
+            exclude: ["group", "longLocal"],
+        }
+    },
+    args: {
+        uploadType: "url",
+        chunkSize: DEFAULT_CHUNK_SIZE,
+        forgetOnSuccess: false,
+        resumeStorage: true,
+        ignoreModifiedDateInStorage: false,
+        sendDataOnCreate: false,
+        sendWithCustomHeader: false,
+    },
+    argTypes: {
+        uploadType: {
+            control: "radio",
+            options: getTusDestinationOptions(),
+        },
+        chunkSize: { control: "number" },
+        forgetOnSuccess: { control: "boolean" },
+        resumeStorage: { control: "boolean" },
+        ignoreModifiedDateInStorage: { control: "boolean" },
+        sendDataOnCreate: { control: "boolean" },
+        sendWithCustomHeader: { control: "boolean" },
     }
-
-    return <TusUploady
-        debug
-        destination={destination}
-        enhancer={composeEnhancers(enhancer, retryEnhancer)}
-        chunkSize={chunkSize}
-        forgetOnSuccess={forgetOnSuccess}
-        resume={resume}
-        ignoreModifiedDateInStorage={ignoreModifiedDateInStorage}
-        sendDataOnCreate={sendDataOnCreate}
-        resumeHeaders={{
-            "x-test-resume": "123",
-            "x-test-override": "abc",
-        }}
-    >
-        <UploadButton id="upload-button">Upload with TUS</UploadButton>
-        <br/>
-        <AbortButton/>
-        <RetryTus/>
-        <ItemProgress/>
-        <ResumeHandler cancelResume={extOptions?.tusCancelResume}/>
-    </TusUploady>;
-};
-
-const tusUploadyStories: CsfExport = getCsfExport(TusUploady, "Tus Uploady", readme, { pkg: "tus-uploady", section: "UI" });
+});
 
 export default { ...tusUploadyStories, title: "UI/Tus Uploady" };